Support more than INT_MAX non-zero elements in sparse matrices.
This change allows ceres to solve larger sparse problems.

It is an intermediate step towards 64bit index support for sparse matrices (which would allow ceres to solve even larger problems and also allow for GPU acceleration in SuiteSparse).

This change does not switch to 64bit cell indices though, so the matrix size is still limited to INT_MAX elements per dimension.

However, values used to index into value and row/col arrays are promoted to 64-bit signed integers, so sparse matrices (except for CompressedRowSparseMatrix) can now have more than INT_MAX nonzero elements.

+++ internal/ceres/block_sparse_matrix.h
@@ -82,7 +82,7 @@ class CERES_EXPORT_INTERNAL BlockSparseMatrix : public SparseMatrix {
   // clang-format off
   int num_rows()         const final { return num_rows_;     }
   int num_cols()         const final { return num_cols_;     }
-  int num_nonzeros()     const final { return num_nonzeros_; }
+  int64_t num_nonzeros() const final { return num_nonzeros_; }
   const double* values() const final { return values_.get(); }
   double* mutable_values()     final { return values_.get(); }
   // clang-format on
@@ -130,8 +130,8 @@ class CERES_EXPORT_INTERNAL BlockSparseMatrix : public SparseMatrix {
  private:
   int num_rows_;
   int num_cols_;
-  int num_nonzeros_;
-  int max_num_nonzeros_;
+  int64_t num_nonzeros_;
+  int64_t max_num_nonzeros_;
   std::unique_ptr<double[]> values_;
   std::unique_ptr<CompressedRowBlockStructure> block_structure_;
 };
