+++ internal/ceres/program.cc
@@ -439,8 +439,9 @@ Program::CreateJacobianBlockSparsityTranspose(int start_residual_block) const {
   const int num_cols = NumResidualBlocks() - start_residual_block;
 
   std::unique_ptr<TripletSparseMatrix> tsm(
-      new TripletSparseMatrix(num_rows, num_cols, 10 * num_cols));
-  int num_nonzeros = 0;
+      new TripletSparseMatrix(num_rows, num_cols,
+                              10 * static_cast<int64_t>(num_cols)));
+  int64_t num_nonzeros = 0;
   int* rows = tsm->mutable_rows();
   int* cols = tsm->mutable_cols();
   double* values = tsm->mutable_values();
